Push initializers out of :android-app
Moving them deeper into their respective modules
