+++ android-app/app/src/main/java/app/tivi/inject/AndroidApplicationComponent.kt
@@ -8,20 +8,17 @@ import android.content.Context
 import app.tivi.BuildConfig
 import app.tivi.TiviApplication
 import app.tivi.app.ApplicationInfo
+import app.tivi.app.Flavor
 import app.tivi.appinitializers.AppInitializer
 import app.tivi.appinitializers.AppInitializers
 import app.tivi.appinitializers.EmojiInitializer
-import app.tivi.appinitializers.LoggerInitializer
-import app.tivi.appinitializers.TmdbInitializer
 import app.tivi.common.imageloading.ImageLoadingComponent
 import app.tivi.data.traktauth.TraktOAuthInfo
 import app.tivi.home.ContentViewSetterComponent
 import app.tivi.tasks.TiviWorkerFactory
 import app.tivi.tmdb.TmdbOAuthInfo
-import app.tivi.util.AppCoroutineDispatchers
 import java.io.File
 import java.util.concurrent.TimeUnit
-import kotlinx.coroutines.Dispatchers
 import me.tatarka.inject.annotations.Component
 import me.tatarka.inject.annotations.IntoSet
 import me.tatarka.inject.annotations.Provides
@@ -44,22 +41,16 @@ abstract class AndroidApplicationComponent(
     abstract val initializers: AppInitializers
     abstract val workerFactory: TiviWorkerFactory
 
-    /**
-     * Need to wait to upgrade to Coroutines 1.7.x so we can reference IO from common
-     */
     @ApplicationScope
     @Provides
-    fun provideCoroutineDispatchers(): AppCoroutineDispatchers = AppCoroutineDispatchers(
-        io = Dispatchers.IO,
-        computation = Dispatchers.Default,
-        main = Dispatchers.Main,
+    fun provideApplicationId(application: Application): ApplicationInfo = ApplicationInfo(
+        packageName = application.packageName,
+        debugBuild = BuildConfig.DEBUG,
+        flavor = when (BuildConfig.FLAVOR) {
+            "qa" -> Flavor.Qa
+            else -> Flavor.Standard
+        },
     )
-
-    @ApplicationScope
-    @Provides
-    fun provideApplicationId(application: Application): ApplicationInfo {
-        return ApplicationInfo(application.packageName)
-    }
 
     @ApplicationScope
     @Provides
@@ -78,14 +69,6 @@ abstract class AndroidApplicationComponent(
     @Provides
     @IntoSet
     fun provideEmojiInitializer(bind: EmojiInitializer): AppInitializer = bind
-
-    @Provides
-    @IntoSet
-    fun provideLoggerInitializer(bind: LoggerInitializer): AppInitializer = bind
-
-    @Provides
-    @IntoSet
-    fun provideTmdbInitializer(bind: TmdbInitializer): AppInitializer = bind
 
     @ApplicationScope
     @Provides
