--- core/logging/src/commonMain/kotlin/app/tivi/util/LoggerComponent.kt
@@ -1,6 +1,0 @@
-// Copyright 2022, Christopher Banes and the Tivi project contributors
-// SPDX-License-Identifier: Apache-2.0
-
-package app.tivi.util
-
-expect interface LoggerComponent
