+++ core/logging/src/iosMain/kotlin/app/tivi/util/LoggerPlatformComponent.kt
@@ -1,4 +1,4 @@
-// Copyright 2022, Google LLC, Christopher Banes and the Tivi project contributors
+// Copyright 2023, Christopher Banes and the Tivi project contributors
 // SPDX-License-Identifier: Apache-2.0
 
 package app.tivi.util
@@ -6,8 +6,8 @@ package app.tivi.util
 import app.tivi.inject.ApplicationScope
 import me.tatarka.inject.annotations.Provides
 
-actual interface LoggerComponent {
+actual interface LoggerPlatformComponent {
     @ApplicationScope
     @Provides
-    fun provideLogger(bind: TimberLogger): Logger = bind
+    fun provideLogger(bind: KermitLogger): Logger = bind
 }
