+++ core/logging/src/jvmMain/kotlin/app/tivi/util/LoggerPlatformComponent.kt
@@ -6,7 +6,7 @@ package app.tivi.util
 import app.tivi.inject.ApplicationScope
 import me.tatarka.inject.annotations.Provides
 
-actual interface LoggerComponent {
+actual interface LoggerPlatformComponent {
     @ApplicationScope
     @Provides
     fun provideLogger(bind: KermitLogger): Logger = bind
